- NEW: The protein translation (forward, reading frame 1 only) is now shown graphically. - NEW: Added new colour schemes to handle the protein stuff. - NEW: Added another scrollable canvas for the protein translation.

+++ src/tablet/gui/viewer/colors/ColorScheme.java
@@ -17,9 +17,9 @@ public abstract class ColorScheme
 		this.contig = contig;
 	}
 
-	public abstract BufferedImage getImage(byte data);
+	public abstract BufferedImage getImage(int data);
 
-	public abstract BufferedImage getConsensusImage(byte data);
+	public abstract BufferedImage getConsensusImage(int data);
 
-	public abstract Color getColor(byte data);
+	public abstract Color getColor(int data);
 }
